Project upgrade to .Net Standard 2.0 SDK style, format nuspec, asmInfo
AssemblyInfo information moved to the project file during the upgrade.

Formatted nuspec file.

Added a basic project icon that looks like an object being cloned.
Add the readme to the NuGet package.
Update the target framework to netstandard2.0.
Moved all the nuget package information to the project file.
Updated copyright.
Version bump to 1.0.2.
Added release notes.
Changed the license to a license expression for the CPL version 1,
 based on the SPDX license expression for that license.
Updated the project URL and repository URL.
Add configs to always build symbol packages, and to use the snupkg
 format.