Fix warnings, formatting, convert from NUnit 2.x to 4.x, net8.0 upgrade
Upgraded the test project to .Net 8.0.
This involved using the Upgrade Assistant and manually working on the
 new SDK-style project file to get everything right.
Installed a number of new packages since with .Net and Core/Standard,
 things are not so built-in.
This is installation of the NUnit Visual Studio Test Context provider
 and the related MS libraries to support using Test Explorer.
This converted it to Package Reference so packages.configh is now gone.
Updated copyright, Target Framework, and platforms.

Updated all NUnit assertions to Constraint-based assertions, except
 for CollectionAssertions which use the Legacy namespace.

Fixed warnings about uninitialized fields with constructors.